+++ src/components/common/Pagination/index.tsx
@@ -1,4 +1,4 @@
-import Button from '@/components/Button';
+import Button from '@/components/common/Button';
 
 export default function Pagination({
   page,
