+++ src/components/pages/main/ActionBar/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '@/components/Button';
+import Button from '@/components/common/Button';
 import useModal from '@/hooks/useModal';
 
 export default function ActionBar({
