+++ src/components/pages/main/BookCard/index.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import Button from '@/components/Button';
+import Button from '@/components/common/Button';
 import useModal from '@/hooks/useModal';
 import { Book } from '@/types/books';
