+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import ModalProvider from '@/components/Modal';
+import ModalProvider from '@/components/common/Modal';
 
 const queryClient = new QueryClient();
